rename the helper function and merge test helpers
